value_set_abstract_objectt correctly evaluates n-ary expressions
expression_transform correctly creates all combinations of operands, but
passed the unmodified expression down for further evaluation. If the
expression was something like
  2 + p
then at the next stage down, we'd look up symbol 'p' and be handed back
a value set. The evaluated would then try to convert that value set to a
constant which would, naturally, fail. Consequently, the expression
would evaluate to top.
Now, we rewrite the expression in terms of the evaluated operands. If
expression was 2 + p, and p was the set { 2, 3 } the expression is
rewritten first as 2 + 2, then as 2 + 3, which we can gather up into our
result of { 4, 5 }.

The various regression tests in this commit exercise various
combinations of constants and value-sets, value-sets of different size,
and so on. All the tests use addition, but the operator is immaterial.