Start making index_ranget more container-like
The index_ranget has, by C++ standards, an unusual interface more akin
to something you'd get on a C# enumerator. This is the first step to
making it more container-like, with begin()/end() member functions
returning iterators.

With this commit, the iterators are input-iterator like, in that they
can only be traversed once.