index_ranget iterators are now forward iterators
index_ranget is now only returns its begin and end
iterators. It's now returned as a value, rather than through a
shared_ptr, as its lifetime will naturally be scope limited and,
as described below, polymorphic behaviour is hidden inside. This
makes is more natural to use too, as we can just write

   for (const auto &index : range)

rather than having to dereference it.

It can now be repeated iterated, rather than the one
and done it was in the previous commit.

Behind it, the classes that do the actual work of enumerating the
ranges have been split out from input_ranget. There is no longer any
need to shared these objects, so they are instantiated using unique_ptr
rather than shared_ptr.