Index expression should used signed_size_type rather than size_type
I based my code in the previous commit on this code, but then found it
caused a type error. This leads me to believe that while this current
code path may not be called in normal usage, it is in fact a latent bug.
I believe this change is correct.